Extract `ECPColorDetection` to own directory.
