+++ libraries/EmbeddedChessPieces/src/ECPColorDetection/ECPColorDetection.cpp
@@ -1,4 +1,4 @@
-#include "EmbeddedChessPieces.h"
+#include "ECPColorDetection.h"
 
 double ECPColorDetection::isWhiteField() {
     double ambient = dezibot.colorSensor.getNormalizedAmbientValue();
