+++ libraries/EmbeddedChessPieces/src/EmbeddedChessPieces.h
@@ -12,47 +12,6 @@
 #ifndef EmbeddedChessPieces_h
 #define EmbeddedChessPieces_h
 
-#include <Dezibot.h>
-
-#define IS_WHITE_FIELD_THRESHOLD 100.0
-
-#define COLOR_CORRECTION_LIGHT_R 43
-#define COLOR_CORRECTION_LIGHT_G 33
-#define COLOR_CORRECTION_LIGHT_B 35
-
-class ECPColorDetection {
-protected:
-    Dezibot dezibot;
-
-public:
-    ECPColorDetection(Dezibot &d) : dezibot(d) {};
-
-    /**
-     * @brief Determine if brightness value represents a white or a black chess
-     *        field.
-     *
-     * Note that the room must be well-lit or color correction light is turned on
-     * (see \see turnOnColorCorrectionLight).
-     * At a normalized ambient light of about 10.0 or lower white field will be interpreted as black.
-     * 
-     * @param brightness normalized brightness value
-     * @return true if surface is white-ish
-     * @return false if surface is black-ish
-     */
-    double isWhiteField();
-
-    /**
-     * @brief Turn on the LED on the bottom of the Dezibot.
-     * 
-     * The light should help the identification of the field color even in a dimly-lit room.
-     */
-    void turnOnColorCorrectionLight();
-
-    /**
-     * @brief Turn off the LED on the bottom of the Dezibot.
-     * 
-     */
-    void turnOffColorCorrectionLight();
-};
+#include "ECPColorDetection/ECPColorDetection.h"
 
 #endif // EmbeddedChessPieces_h
